requirements.txt
asgiref==3.4.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Django==4.0.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
filelock==3.6.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.4.0
hyperframe==5.2.0
idna==2.10
joblib==1.1.0
langdetect==1.0.9
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
rake-nltk==1.0.6
regex==2022.3.15
requests==2.27.1
rfc3986==1.5.0
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
sentencepiece==0.1.96
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.8.2+cpu
torchaudio==0.8.2
torchvision==0.9.2+cpu
tqdm==4.63.0
transformers==4.17.0
typing-extensions==4.1.1
tzdata==2021.5
urllib3==1.26.9
